ENH: add pages and navigation
